+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Downgraded CMake to version 3.25 to improve compatibility.
 - Added `--type` option for filtering by type.
 - Added tests for `dmi_date_parse()`.
+- Fixed `dmi_file_read()` behaviour on SysFS.
+- Fixed `dmi_file_map()` error handling.
 - Fixed firmware ROM size decoding.
 - Fixed platform firmware version decoding.
 - Fixed embedded controller firmware version decoding.